Chapter 2
I like the concept of Data Transfer Objects, especially with the help of
mapped-types to make them reusable for different verbs.
I like the strictness of how Nest applications handle the requests
thanks to this, but I also wonder how many files and imports will need
to be present in a production application.
Homework will likely be to dig through the code of some other
applications and see how they use DTOs.